requirements.txt
langchain==0.1.1
python-dotenv==1.0.0
streamlit==1.29.
langchain_pinecone
sentence_transformers